+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v0.34.2] - 2023-12-18
+### Changed
+* **Washer/WasherDryer:** Added undocumented option value `LaundryCare.Washer.EnumType.Temperature.Auto`. Added several other undocumented program options, with guesses at their possible values. 
+### Fixed
+* **Dishwasher/Dryer/Oven/Washer/WasherDryer:** Corrected expected type for the `BSH.Common.Option.RemainingProgramTimeIsEstimated` program option. (#187)
+
 ## [v0.34.1] - 2023-12-18
 ### Changed
 * Tweaked the messages that are written to the log when the supported options cannot be read for some programs. This is because some appliances advertise support for programs that cannot be selected. (#186)
@@ -473,7 +479,8 @@ All notable changes to this project will be documented in this file.
 
 Copyright © 2019-2023 Alexander Thoukydides
 
-[Unreleased]:       https://github.com/thoukydides/homebridge-homeconnect/compare/v0.34.1...HEAD
+[Unreleased]:       https://github.com/thoukydides/homebridge-homeconnect/compare/v0.34.2...HEAD
+[v0.34.2]:          https://github.com/thoukydides/homebridge-homeconnect/compare/v0.34.1...v0.34.2
 [v0.34.1]:          https://github.com/thoukydides/homebridge-homeconnect/compare/v0.34.0...v0.34.1
 [v0.34.0]:          https://github.com/thoukydides/homebridge-homeconnect/compare/v0.33.1...v0.34.0
 [v0.33.1]:          https://github.com/thoukydides/homebridge-homeconnect/compare/v0.33.0...v0.33.1